Balance converted to int - Oryx

+++ app/Http/Controllers/OryxGamingController.php
@@ -638,7 +638,7 @@ class OryxGamingController extends Controller
 
 	private function _toPennies($value)
 	{
-	    return str_replace(' ', '', intval(
+	    return (int) str_replace(' ', '', intval(
 	        strval(floatval(
 	            preg_replace("/[^0-9.]/", "", $value)
 	        ) * 100)
@@ -647,7 +647,7 @@ class OryxGamingController extends Controller
 
 	private function _toDollars($value)
 	{
-		return str_replace(' ', '', number_format(($value / 100), 2, '.', ' '));
+		return (int) str_replace(' ', '', number_format(($value / 100), 2, '.', ' '));
 	}
 
 }
